[refactor] UtilsTemplate를 생성해서 각 도메인에 대한 Test Template에 공통으로 들어가는 로직을 모아둠
