+++ src/test/java/springJpaBoard/Board/board/BoardTemplate.java
@@ -1,9 +1,6 @@
 package springJpaBoard.Board.board;
 
 import org.jetbrains.annotations.NotNull;
-import springJpaBoard.Board.domain.Address;
-import springJpaBoard.Board.domain.Board;
-import springJpaBoard.Board.domain.Member;
 
 import static springJpaBoard.Board.controller.boarddto.BoardDto.ModifyBoardRequest;
 import static springJpaBoard.Board.controller.boarddto.BoardDto.ModifyBoardResponse;
@@ -25,29 +22,6 @@ public class BoardTemplate {
                 .title("2")
                 .writer("writer")
                 .content("2")
-                .build();
-    }
-
-
-    @NotNull
-    public static Board getBoard() {
-        return Board.builder()
-                .title("title")
-                .writer("writer")
-                .content("content")
-                .build();
-    }
-
-
-    @NotNull
-    public static Member getMember() {
-        return Member.builder()
-                .id(1L)
-                .name("1")
-                .gender("남성")
-                .loginId("1")
-                .password("1")
-                .address(new Address("1", "1", "1"))
                 .build();
     }
 }
