+++ src/test/java/springJpaBoard/Board/board/controller/BoardApiControllerTest.java
@@ -29,6 +29,7 @@ import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.BDDMockito.given;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.*;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;
+import static springJpaBoard.Board.UtilsTemplate.*;
 import static springJpaBoard.Board.board.BoardTemplate.*;
 import static springJpaBoard.Board.controller.boarddto.BoardDto.*;
