+++ src/test/java/springJpaBoard/Board/comment/controller/CommentApiControllerTest.java
@@ -24,6 +24,7 @@ import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.BDDMockito.given;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;
+import static springJpaBoard.Board.UtilsTemplate.*;
 import static springJpaBoard.Board.comment.CommentTemplate.*;
 
 @ExtendWith(SpringExtension.class)
